+++ src/Traits/CounterpartyProperty.php
@@ -13,44 +13,52 @@ trait CounterpartyProperty
      * Установка типа контрагента.
      *
      * @param  string  $counterpartyType  Тип контрагента (PrivatePerson)
-     * @return void
+     * @return $this
      */
-    public function setCounterpartyType(string $counterpartyType): void
+    public function setCounterpartyType(string $counterpartyType): self
     {
         $this->counterpartyType = $counterpartyType;
+
+        return $this;
     }
 
     /**
      * Установка свойства контрагента.
      *
      * @param  string  $counterpartyProperty  Свойства контрагента (Recipient / Sender)
-     * @return void
+     * @return $this
      */
-    public function setCounterpartyProperty(string $counterpartyProperty): void
+    public function setCounterpartyProperty(string $counterpartyProperty): self
     {
         $this->counterpartyProperty = $counterpartyProperty;
+
+        return $this;
     }
 
     /**
      * Установка формы собственности.
      *
      * @param  string  $ownershipForm  Ref формы собственности
-     * @return void
+     * @return $this
      */
-    public function setOwnershipForm(string $ownershipForm): void
+    public function setOwnershipForm(string $ownershipForm): self
     {
         $this->ownershipForm = $ownershipForm;
+
+        return $this;
     }
 
     /**
      * Установка ЕДРПОУ.
      *
      * @param  string  $EDRPOU  ЕРДПОУ
-     * @return void
+     * @return $this
      */
-    public function setEDRPOU(string $EDRPOU): void
+    public function setEDRPOU(string $EDRPOU): self
     {
         $this->EDRPOU = $EDRPOU;
+
+        return $this;
     }
 
     /**
@@ -111,7 +119,7 @@ trait CounterpartyProperty
      * Вынесенная логика для удаления контактов
      * используется для создания / изменения данных организаций или третьих лиц.
      *
-     * @deprecated НЕ СДЕЛАНО
+     * @deprecated НЕ СДЕЛАНО, сложно проверить
      *
      * TODO Не сделано
      *